refactor: construct memory package
